refactor to have Itinerary inside ItineraryWrapper
